Empty object checking and db function typing

+++ controllers/userController.ts
@@ -5,15 +5,17 @@ import {
   updateUser,
   deleteUser,
   getUserByUsername,
+  getUserById,
   getUserAndUpdateUser,
 } from "../models/modelsMod.ts";
 import { generateJWT } from "../auth/authMod.ts";
+import { isEmptyObject } from "../utils/utilsMod.ts";
 
 export async function loginHandler(ctx: RouterContext<string>) {
   const { username } = await ctx.request.body.json();
   try {
     const user = getUserByUsername(username);
-    if (!user) {
+    if (isEmptyObject(user)) {
       ctx.response.status = 401;
       ctx.response.body = { error: "Invalid username or password" };
       return;
@@ -60,9 +62,12 @@ export function getSingleUserHandler(ctx: RouterContext<string>) {
   const username = ctx.params.username;
   try {
     const user = getUserByUsername(username);
-    //  The response body is destructured as { user } to create a properly structured JSON response
-    //  with a named "user" property, following REST API conventions and making the response
-    //  more explicit for clients consuming the API.
+    console.log(user);
+    if (isEmptyObject(user)) {
+      ctx.response.status = 404;
+      ctx.response.body = { error: "User not found" };
+      return;
+    }
     ctx.response.body = { user };
   } catch (error) {
     console.log(error);
@@ -76,6 +81,14 @@ export async function userUpdaterHandler(ctx: RouterContext<string>) {
   //   const { username, email, userId } = await ctx.request.body.json();
   const { username, email } = await ctx.request.body.json();
   try {
+    const userDoesNotExist = getUserByUsername(username);
+    if (isEmptyObject(userDoesNotExist)) {
+      ctx.response.status = 404;
+      ctx.response.body = {
+        error: "The user you want to update was not found",
+      };
+      return;
+    }
     updateUser(userId, username, email);
     ctx.response.status = 200;
     ctx.response.body = {
@@ -92,6 +105,14 @@ export function userDeleterHandler(ctx: RouterContext<string>) {
   //   const userId = +ctx.params.userId;
   const userId = Number(ctx.params.userId);
   try {
+    const userDoesNotExist = getUserById(userId);
+    if (isEmptyObject(userDoesNotExist)) {
+      ctx.response.status = 404;
+      ctx.response.body = {
+        error: "The user you want to delete was not found",
+      };
+      return;
+    }
     deleteUser(userId);
     ctx.response.status = 200;
     ctx.response.body = {
@@ -109,6 +130,14 @@ export async function getUserAndUpdateHandler(ctx: RouterContext<string>) {
   const { email, username } = await ctx.request.body.json();
   try {
     const updatedUser = getUserAndUpdateUser(userId, email, username);
+    const userDoesNotExist = getUserByUsername(username);
+    if (isEmptyObject(userDoesNotExist)) {
+      ctx.response.status = 404;
+      ctx.response.body = {
+        error: "The user you want to update was not found",
+      };
+      return;
+    }
     ctx.response.status = 200;
     ctx.response.body = {
       updatedUser,
